+++ curve25519-dalek/src/lib.rs
@@ -98,7 +98,10 @@ pub(crate) mod window;
 mod lizard;
 
 pub use crate::{
-    edwards::EdwardsPoint, montgomery::MontgomeryPoint, ristretto::RistrettoPoint, scalar::Scalar,
+    edwards::EdwardsPoint,
+    montgomery::MontgomeryPoint,
+    ristretto::RistrettoPoint,
+    scalar::{HalfWidthScalar, Scalar},
 };
 
 // Build time diagnostics for validation