+++ curve25519-dalek/CHANGELOG.md
@@ -5,6 +5,11 @@ major series.
 
 # 5.x series
 
+## Unreleased
+
+* Add `EdwardsPoint::vartime_triple_scalar_mul_basepoint`, computing `a1*A1 + a2*A2 + b*B` in roughly half the doublings of the naive approach when `a1` and `a2` are less than 2^128 ([#858](https://github.com/dalek-cryptography/curve25519-dalek/pull/858))
+* Add `HalfWidthScalar`, a `Scalar` that is known to be less than 2^128 ([#858](https://github.com/dalek-cryptography/curve25519-dalek/pull/858))
+
 ## 5.0.0 - 2026-07-06
 
 ### Breaking Changes